--- a/Rock.Migrations/Rock.Migrations.csproj
+++ b/Rock.Migrations/Rock.Migrations.csproj
@@ -74,6 +74,11 @@
     <Compile Include="Migrations\Version 6.0\Version 1.6.0\201609221415156_PreventDupCheckin.Designer.cs">
       <DependentUpon>201609221415156_PreventDupCheckin.cs</DependentUpon>
     </Compile>
+    <Compile Include="Migrations\201609140843400_PhoneNumberReversed.cs" />
+    <Compile Include="Migrations\Version 5.0\Version 1.5.0\201606061308109_ClearPreviousMigrations5.cs" />
+    <Compile Include="Migrations\Version 5.0\Version 1.5.0\201606061308109_ClearPreviousMigrations5.Designer.cs">
+      <DependentUpon>201606061308109_ClearPreviousMigrations5.cs</DependentUpon>
+    </Compile>
     <Compile Include="Migrations\Version 6.0\Version 1.6.0\201609221633157_BodyPageCSS-ContentChannelNoContent.cs" />
     <Compile Include="Migrations\Version 6.0\Version 1.6.0\201609221633157_BodyPageCSS-ContentChannelNoContent.Designer.cs">
       <DependentUpon>201609221633157_BodyPageCSS-ContentChannelNoContent.cs</DependentUpon>
@@ -2287,6 +2292,13 @@
     </EmbeddedResource>
     <EmbeddedResource Include="Migrations\Version 6.0\Version 1.6.0\201606232056332_WorkflowNumber.resx">
       <DependentUpon>201606232056332_WorkflowNumber.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201609140843400_PhoneNumberReversed.resx">
+      <DependentUpon>201609140843400_PhoneNumberReversed.cs</DependentUpon>
+      <SubType>Designer</SubType>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\Version 5.0\Version 1.5.0\201606061308109_ClearPreviousMigrations5.resx">
+      <DependentUpon>201606061308109_ClearPreviousMigrations5.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Migrations\Version 6.0\Version 1.6.0\201606241700105_ExceptionUpdates.resx">
       <DependentUpon>201606241700105_ExceptionUpdates.cs</DependentUpon>