--- a/Rock.Migrations/Rock.Migrations.csproj
+++ b/Rock.Migrations/Rock.Migrations.csproj
@@ -78,10 +78,6 @@
 	 <Compile Include="Migrations\201609140843400_PhoneNumberReversed.Designer.cs">
       <DependentUpon>201609140843400_PhoneNumberReversed.cs</DependentUpon>
     </Compile>
-    <Compile Include="Migrations\Version 5.0\Version 1.5.0\201606061308109_ClearPreviousMigrations5.cs" />
-    <Compile Include="Migrations\Version 5.0\Version 1.5.0\201606061308109_ClearPreviousMigrations5.Designer.cs">
-      <DependentUpon>201606061308109_ClearPreviousMigrations5.cs</DependentUpon>
-    </Compile>
     <Compile Include="Migrations\Version 6.0\Version 1.6.0\201609221633157_BodyPageCSS-ContentChannelNoContent.cs" />
     <Compile Include="Migrations\Version 6.0\Version 1.6.0\201609221633157_BodyPageCSS-ContentChannelNoContent.Designer.cs">
       <DependentUpon>201609221633157_BodyPageCSS-ContentChannelNoContent.cs</DependentUpon>
@@ -2299,9 +2295,6 @@
     <EmbeddedResource Include="Migrations\201609140843400_PhoneNumberReversed.resx">
       <DependentUpon>201609140843400_PhoneNumberReversed.cs</DependentUpon>
       <SubType>Designer</SubType>
-    </EmbeddedResource>
-    <EmbeddedResource Include="Migrations\Version 5.0\Version 1.5.0\201606061308109_ClearPreviousMigrations5.resx">
-      <DependentUpon>201606061308109_ClearPreviousMigrations5.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Migrations\Version 6.0\Version 1.6.0\201606241700105_ExceptionUpdates.resx">
       <DependentUpon>201606241700105_ExceptionUpdates.cs</DependentUpon>