--- a/Rock/Workflow/Action/CheckIn/FindFamilies.cs
+++ b/Rock/Workflow/Action/CheckIn/FindFamilies.cs
@@ -54,8 +54,10 @@ namespace Rock.Workflow.Action.CheckIn
             {
                 var personService = new PersonService( rockContext );
                 var memberService = new GroupMemberService( rockContext );
+                GroupService groupService = new GroupService( rockContext );
+                PhoneNumberService phoneNumberService = new PhoneNumberService( rockContext );
 
-                Guid familyGroupTypeGuid = SystemGuid.GroupType.GROUPTYPE_FAMILY.AsGuid();
+                int familyGroupTypeId = GroupTypeCache.Read( Rock.SystemGuid.GroupType.GROUPTYPE_FAMILY.AsGuid() ).Id;
                 var dvActive = DefinedValueCache.Read( Rock.SystemGuid.DefinedValue.PERSON_RECORD_STATUS_ACTIVE.AsGuid() );
 
                 if ( checkInState.CheckIn.SearchType.Guid.Equals( new Guid( SystemGuid.DefinedValue.CHECKIN_SEARCH_TYPE_PHONE_NUMBER ) ) )
@@ -65,31 +67,48 @@ namespace Rock.Workflow.Action.CheckIn
                     var personRecordTypeId = DefinedValueCache.Read( Rock.SystemGuid.DefinedValue.PERSON_RECORD_TYPE_PERSON.AsGuid() ).Id;
 
                     // Find the families with any member who has a phone number that contains selected value
-                    var familyQry = memberService
-                        .Queryable().AsNoTracking()
-                        .Where( m =>
-                            m.Group.GroupType.Guid.Equals( familyGroupTypeGuid ) &&
-                            m.Person.RecordTypeValueId == personRecordTypeId );
+                    var familyQry = phoneNumberService.Queryable().AsNoTracking();
 
-                    if ( checkInState.CheckInType != null && checkInState.CheckInType.PreventInactivePeopele && dvActive != null )
-                    {
-                        familyQry = familyQry.Where( m =>
-                            m.Person.RecordStatusValueId == dvActive.Id );
-                    }
+
 
                     if ( checkInState.CheckInType == null || checkInState.CheckInType.PhoneSearchType == PhoneSearchType.EndsWith )
                     {
-                        familyQry = familyQry.Where( m =>
-                            m.Person.PhoneNumbers.Any( n => n.Number.EndsWith( numericPhone ) ) );
+                        char[] charArray = numericPhone.ToCharArray();
+                        Array.Reverse( charArray );
+                        familyQry = familyQry.Where( o =>
+                            o.NumberReversed.StartsWith( new string( charArray ) ) );
                     }
                     else
                     {
-                        familyQry = familyQry.Where( m =>
-                            m.Person.PhoneNumbers.Any( n => n.Number.Contains( numericPhone ) ) );
+                        familyQry = familyQry.Where( o =>
+                            o.Number.Contains( numericPhone ) );
                     }
 
-                    var familyIdQry = familyQry
-                        .Select( m => m.GroupId )
+                    var tmpQry = familyQry.Join( personService.Queryable().AsNoTracking(),
+                            o => new { PersonId = o.PersonId, IsDeceased = false, RecordTypeValueId = personRecordTypeId },
+                            p => new { PersonId = p.Id, IsDeceased = p.IsDeceased, RecordTypeValueId = p.RecordTypeValueId.Value },
+                            ( pn, p ) => new { Person = p, PhoneNumber = pn } )
+                            .Join( memberService.Queryable().AsNoTracking(),
+                            pn => pn.Person.Id,
+                            m => m.PersonId,
+                            ( o, m ) => new { PersonNumber = o.PhoneNumber, GroupMember = m } );
+
+
+                    if ( checkInState.CheckInType != null && checkInState.CheckInType.PreventInactivePeopele && dvActive != null )
+                    {
+                        // Reconstruct the query because doing this in the join condition is the fastest way
+                        tmpQry = familyQry.Join( personService.Queryable().AsNoTracking(),
+                            o => new { PersonId = o.PersonId, IsDeceased = false, RecordTypeValueId = personRecordTypeId, RecordStatusValueId = dvActive.Id },
+                            p => new { PersonId = p.Id, IsDeceased = p.IsDeceased, RecordTypeValueId = p.RecordTypeValueId.Value, RecordStatusValueId = p.RecordStatusValueId.Value },
+                            ( pn, p ) => new { Person = p, PhoneNumber = pn } )
+                            .Join( memberService.Queryable().AsNoTracking(),
+                            pn => pn.Person.Id,
+                            m => m.PersonId,
+                            ( o, m ) => new { PersonNumber = o.PhoneNumber, GroupMember = m } );
+                    }
+
+                    var familyIdQry = groupService.Queryable().Where( g => tmpQry.Any( o => o.GroupMember.GroupId == g.Id ) && g.GroupTypeId == familyGroupTypeId )
+                        .Select( g => g.Id )
                         .Distinct();
 
                     int maxResults = checkInState.CheckInType != null ? checkInState.CheckInType.MaxSearchResults : 100;
@@ -157,7 +176,7 @@ namespace Rock.Workflow.Action.CheckIn
 
                     foreach ( var person in people )
                     {
-                        foreach ( var group in person.Members.Where( m => m.Group.GroupType.Guid.Equals( familyGroupTypeGuid ) ).Select( m => m.Group ).ToList() )
+                        foreach ( var group in person.Members.Where( m => m.Group.GroupTypeId == familyGroupTypeId ).Select( m => m.Group ).ToList() )
                         {
                             var family = checkInState.CheckIn.Families.Where( f => f.Group.Id == group.Id ).FirstOrDefault();
                             if ( family == null )