# Context
Phone number searches during check-in were slow and put extensive load on the SQL server.  We were seeing SQL waits  and DB contention that we were pretty uncomfortable with during checkin times.

# Goal
The goal is to speed up the FindFamilies workflow action as much as possible.

# Strategy
This is a 2-fold approach to improving the FindFamilies searching. The majority of the brain-power behind this solution is from @tom-powers who helped optimize the query and is solely responsible for the phone number reversal:

- First, the query itself has been optimized to search the phonenumber table with an inner join on the person and group member to get the family id.  This as as optimized as we could find query-wise. 
- Second, we updated the PhoneNumber model to add a phone number reversed computed column to the table.  This literally stores the phone number backwards in the database, which, when searched using a reversed phonenumber "Starts With" style query, allows for the usage of an index which dramatically improves the results of this search (see the Statistics below).

# Possible Implications
This just needs to be tested well.  I've tested it pretty well on this end but more testing is better!

# Statistics
@trydyingtolive Created a phone number load test block which executed this 10 times every iteration in a loop.  These are examples of the time improvements we saw:
### Before:
> Searched: 4736 in 502ms
> Searched: 4942852021 in 419ms
> Searched: 4648 in 510ms
> Searched: 359599930 in 418ms
> Searched: 8699 in 484ms
> Searched: 4931834012 in 388ms
> Searched: 1751 in 588ms
> Searched: 521065389 in 439ms
> Searched: 1172 in 514ms
> Searched: 850907461 in 416ms

### After
> Searched: 4431 in 70ms
> Searched: 4662344237 in 35ms
> Searched: 6178 in 71ms
> Searched: 1207021813 in 33ms
> Searched: 5392 in 71ms
> Searched: 515678527 in 50ms
> Searched: 5323 in 71ms
> Searched: 1340554413 in 33ms
> Searched: 5241 in 79ms
> Searched: 777506990 in 34ms