@edmistj 2016-10-06
This approach makes a lot of sense. Have you run this code in production?