@cbump 2016-10-06
@edmistj Yes, we are using it in production right now.  We have a custom FindFamilies workflow action that we use (once this PR gets merged we can use the core one again) and there were a couple changes made to FindFamilies for excluding Inactive records a few weeks ago that I merged into this specific PR.