@azturner 2016-10-06
@cbump I pulled this in and am working on redoing the migration since we had already added more migrations since your branch. However I don't see where you were creating the computed column in a migration.  Can you post the properties of that column so I can see what the formula is for it?