@cbump 2016-10-06
@azturner It's a persisted computed column.  If you look at how the column is created the reverse() function is actually in the definition: https://github.com/SparkDevNetwork/Rock/blob/5166f8e3d0c988b4d6b7903a1d188ca050e61be5/Rock.Migrations/Migrations/201609140843400_PhoneNumberReversed.cs#L29